requirements.txt
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
dj-database-url==0.5.0
Django==3.0
django-extensions==2.2.5
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.4
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
whitenoise==4.1.4
